autoscaler/requirements.txt
PyYAML
requests
docker>=6.0.0